Pull pattern matching code out into its own file
