+++ lib/ast/CMakeLists.txt
@@ -1,8 +1,6 @@
-set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -g")
-set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -g")
-
 add_library(AST
   AST.cpp
+  pattern_matching.cpp
 )
 
 target_link_libraries(AST
